@@ -1,4 +1,4 @@
-name: flutter_sns_form
+name: memorybeating
 description: A new Flutter project.
 # The following line prevents the package from being accidentally published to
 # pub.dev using `flutter pub publish`. This is preferred for private packages.
@@ -19,7 +19,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=3.1.1 <4.0.0'
+  sdk: '>=3.1.0 <4.0.0'
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -36,15 +36,10 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
   get: ^4.6.6
-  cached_network_image: ^3.3.0
-  expandable_text: ^2.3.0
-  photo_manager: ^2.7.1
-  camera: ^0.10.5+5
-  path_provider: ^2.1.1
-  path: ^1.8.3
-  image_gallery_saver: ^2.0.3
-  image_picker: ^1.0.4
-  cross_file: ^0.3.3+6
+  flutter_svg: ^2.0.9
+  webview_flutter: ^4.7.0
+  flutter_inappwebview: ^6.0.0
+  camera: ^0.10.5+6
 dev_dependencies:
   flutter_test:
     sdk: flutter
@@ -70,6 +65,10 @@ flutter:
   # To add assets to your application, add an assets section, like this:
   assets:
     - assets/images/
+    - assets/images/profile_logo.svg
+    - assets/images/nft_logo.svg
+    - assets/images/fest_logo.svg
+
   #   - images/a_dot_burr.jpeg
   #   - images/a_dot_ham.jpeg
 
